requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
psutil==5.9.6
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
hdbscan==0.8.33
structlog==23.2.0
# insightface==0.7.3  # Заменено на face_recognition для лучшей точности
# onnxruntime==1.16.3  # Не требуется

# face_recognition (dlib-based) - точность 99.38% на LFW, легковесный
face-recognition==1.3.0
face-recognition-models==0.3.0
dlib==19.24.6

